requirements.txt
bcrypt==3.2.0
certifi==2020.6.20
cffi==1.14.2
cryptography==3.1
future==0.18.2
netmiko==3.3.0
paramiko==2.7.2
progress-bar==8
pycparser==2.20
PyNaCl==1.4.0
pyserial==3.4
scp==0.13.2
six==1.15.0
